Added Calendar Class - Replaces GraduateAvailableTimes class
Attempting to update the Scheduler class and perform initialization,
working with the ArrayList structure proved to be troublesome: values
were not being initialized properly and causing
IndexOutOfBoundsExceptions.  Now, rather than having an ArrayList of
GraduateAvailableTimes, all of the information (days and times) for the
week is contained in the Calendar class.

New methods have been added to the GraduateAssistant class. These
methods allow the availability of the GA to be set/get using the day of
the week, and a specific time or a range of times.

There is a brief example of how the Class class and GraduateAssistant
class might be used together in the FXMLDocumentController, under the
runAlgorithm event