Input data from past quarters
Data from Fall quarter has been populated into data structures.  Running
the test, it is apparent that too many classes are being assigned to GA
0 and GA 1.  Hence, the check for the maximum number of hours is not
working (or rather, the amount of time the GA is working is not being
updated).  Additionally, it's worth noting that with the current setup,
the first intitialized GAs are going to receive classes first. This may
be a conflict.