Alternating path (backtracking) complete. Adding partial assignment
Functionality for the alternating path (backtracking) is completed, and
has been run through the Test_Algorithm_BaseCase.  

Adding functionality for partial assignment.  Currently, the first step
involved ignoring any qualifications while attempting to assign the GA.
Later steps will attempt to assign GAs that are available for part of
the times the classes are offered.